Fixed time calculiteng bag and updated README

+++ GraphProcessorAPI/src/DTOs.cs
@@ -7,7 +7,7 @@
         public string Algorithm { get; set; }
         public string StartVertex { get; set; }
         public List<string> ShortestPath { get; set; }
-        public double TimeNs { get; set; }
+        public long TimeNs { get; set; }
     }
 
     public class DijkstraResultDTO : AlgorithmResultDTO
